requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.

south
django-admin-tools==0.3.0
django-extensions
django-filebrowser==3.1
django-grappelli==2.2

PIL